Implement and setup polyversion script.
The script replaces `sphinx-multiversion` and builds multiple versions
of the documentation which are sourced from the git repo.
The script was created since it supports more features than than
sphinx-multiversion while being less complicated. `sphinx-multiversion` on the
other hand is no longer maintained.

This also updates existing files for use with the new script and
adds some files that will be placed in the root of the polyversion build.

The script depends on `jinja2` which is already an indirekt dependency
through `sphinx`.

* docs/sphinx-polyversion.py : Implement.

* pyproject.toml : Add `jinja2` to group `docs`.

* poetry.lock : Autogenerate through `pre-commit`.

* docs/source/conf.py : Adjust to new script. Fix some more stuff.

* docs/source/_templates/versioning.html : Adjust to new script.

* docs/source/_polyversion/templates/index.html : Added.
    Its a template for the root file of the documentation redirecting
    to the latest tags root.

* docs/source/_polyversion/static/.nojekyll : Added.
    Its a marker file for github-pages.